Refactor dataset/model fetching to lisbet.hub and update records API
- Move fetch_dataset and fetch_model to new lisbet.hub module
- Update CLI to use lisbet.hub for fetch commands
- Refactor datasets to use Record dataclass instead of dicts
- Remove fetch_dataset from datasets.core and update __init__.py
- Update calms21 and mabe22 to return lists of Record objects
