+++ src/lisbet/datasets/__init__.py
@@ -1,4 +1,4 @@
-from .core import dump_records, fetch_dataset, load_records
+from .core import dump_records, load_records
 
-__all__ = ["dump_records", "fetch_dataset", "load_records"]
+__all__ = ["dump_records", "load_records"]
 __doc__ = """Datasets input for LISBET."""
