+++ src/lisbet/datasets/mabe22.py
@@ -27,6 +27,8 @@ import xarray as xr
 from movement.io import load_poses
 from tqdm.auto import tqdm
 
+from .core import Record
+
 
 def _preprocess_mabe22_sequence(raw_positions):
     """Preprocess a sequence in the MABe22 Mouse Triplets dataset."""
@@ -111,9 +113,9 @@ def _load_train(train_path):
         )
 
         # Create record data structure
-        rec_data = {"posetracks": posetracks, "annotations": annotations}
+        record = Record(id=rec_id, posetracks=posetracks, annotations=annotations)
 
-        train_records.append((rec_id, rec_data))
+        train_records.append(record)
 
     return train_records
 
@@ -196,9 +198,9 @@ def _load_test(test_seq_path, test_labels_path):
         annotations = xr.merge([annot_cls, annot_reg])
 
         # Create record data structure
-        rec_data = {"posetracks": posetracks, "annotations": annotations}
+        record = Record(id=rec_id, posetracks=posetracks, annotations=annotations)
 
-        test_records.append((rec_id, rec_data))
+        test_records.append(record)
 
     return test_records
